moved chess to standalone folder
