+++ client/src/chess/components/Board.tsx
