+++ client/src/chess/components/PlayerDisplay.tsx
