+++ client/src/chess/components/Square.tsx
