+++ client/src/chess/game/Pieces/Bishop.ts
