+++ client/src/chess/game/Pieces/Queen.ts
