+++ client/src/chess/game/Pieces/index.ts
