+++ client/src/chess/types/game.d.ts
