+++ client/src/pages/Chess.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import NavBar from '../components/Page/NavBar';
 import PageContent from '../components/Page/PageContent';
 import Footer from '../components/Page/Footer';
-import Board from '../components/Chess/Board';
+import Board from '../chess/Board';
 import { Text, Box, Flex, Link } from "@chakra-ui/react"
 import { Link as ReactLink } from "react-router-dom";
