+++ client/src/react-app-env.d.ts
@@ -1,9 +1,9 @@
 /// <reference types="react-scripts" />
 /// <reference path="./pages/Home/Projects/SearchFilter.d.ts" />
 /// <reference path="./graphql/types.d.ts" />
-/// <reference path="./components/Chess/types/game.d.ts"/>
-/// <reference path="./components/Chess/types/context.d.ts"/>
-/// <reference path="./components/Chess/types/components.d.ts"/>
+/// <reference path="./chess/types/game.d.ts"/>
+/// <reference path="./chess/types/context.d.ts"/>
+/// <reference path="./chess/types/components.d.ts"/>
 
 declare module "react-syntax-highlighter"
 declare module "react-syntax-highlighter/dist/esm/styles/hljs"
